Add Documentation to ExecSorts File and Migrate CleanAST Function into ExecSorts
